+++ example/example.dart
@@ -1,3 +1,4 @@
+import 'dart:convert';
 import 'dart:io';
 import 'package:alexaquery_dart/alexaquery_dart.dart';
 
@@ -7,6 +8,12 @@ void main() async {
     logger: (log, level) => print("$level: $log"),
     loginToken: "",
   );
-  var devices = await alexaQuery.getDevices("testuser");
-  print(devices);
+  // var devices = await alexaQuery.getDevices("testuser");
+  // print(devices);
+
+  // final memories = await alexaQuery.getMemories("testuser");
+  // print(jsonEncode(memories));
+
+  final queue = await alexaQuery.getQueue("testuser", "Bedroom Pop");
+  print(jsonEncode(queue));
 }